+++ modules/Users/Search.inc.php
@@ -79,7 +79,7 @@ if ( empty( $_REQUEST['search_modfunc'] ) )
 			echo '<label><input type="checkbox" name="include_inactive" value="Y" /> ' .
 			_( 'Include Parents of Inactive Students' ) . '</label><br />';
 
-			echo '<br />' . Buttons( _( 'Submit' ), _( 'Reset' ) ) . '</div><br />';
+			echo '<br />' . Buttons( _( 'Submit' ) ) . '</div><br />';
 
 			if ( ! empty( $extra['search'] )
 				|| ! empty( $extra['extra_search'] )